update document process update graph

+++ src/main/java/edu/fudan/backend/dao/ImageRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
 public interface ImageRepository extends JpaRepository<Image, Long> {
-    List<Image> findAllByTopicId(Long topicId);
+    List<Image> findAllByTopicId(Integer topicId);
 }
